Update client side setup tool and schedule job to 2.0.0
* Update client side scheduled linux job to 2.0.0
  * Change to parsing the .duplicacy/preferences file to determine things like CLIENT SERVER and SERVER_PORT instead of setting these in the config.bash file
    * This requires python3 to parse the file
  * Support multiple BACKUP_DIRECTORIES instead of just hardcoded ${DUPLICACY_BASEDIR}/backup
    * As part of this, support both local backup destinations and sftp destinations (previously only a single sftp destination was supported)
    * The tool now iterates over all BACKUP_DIRECTORIES and backs up each
    * Since this now supports multiple destinations, the tool will upload the logs to the *last* destination in the list
  * Support having a different CLIENT (the sftp/ssh username which the client authenticates to the backup destination with ) from CLIENT_INDIVIDUAL_ID (the hostname of the client)
    * This allows multiple clients to use the same CLIENT value (and different CLIENT_INDIVIDUAL_ID values) and benefit from deduplication between them
  * Instead of fetching the healthchecks.io HC_UUID from the hc_uuid.clientname file on the destination sftp backup server, the tool now expects to find HC_UUID set in the config.bash
    * This is so that a "start" command can be sent to healthchecks.io earlier in the run (before connecting to the backup destination) to be able to catch failures that wouldn't previously be caught
* Update client side setup-client to 2.0.0
  * Support configuring BACKUP_DIRECTORY in config.bash instead of assuming ${DUPLICACY_BASEDIR}/backup
  * Support new LOCAL_DESTINATION_DIRECTORY setting if the destination is a local directory instead of a remote sftp server
  * Allow for multiple runs
    * Only update the KNOWN_HOSTS file if something has changed
    * Only update the systemd service and timer or the cron file if the files don't exist (don't overwrite the files that are already there)
  * Prompts for what the HC_UUID value is so it can be written to config.bash
  * Changes from a single BACKUP_DIRECTORY to a list of BACKUP_DIRECTORIES so you can have multiple backup directories on a single client
  * Add a reminder to remove the RSA private key from the client and store it somewhere safe

+++ setup-client.bash
@@ -1,8 +1,8 @@
 #!/bin/bash -e
 
-VERSION=1.0.0
+VERSION=2.0.0
 
-DUPLICACY_URL="https://github.com/gilbertchen/duplicacy/releases/download/v3.1.0/duplicacy_linux_x64_3.1.0"
+DUPLICACY_URL="https://github.com/gilbertchen/duplicacy/releases/download/v3.2.4/duplicacy_linux_x64_3.2.4"
 DUPLICACY_BASEDIR=/opt/duplicacy
 
 message () {
@@ -17,26 +17,35 @@ if [ ! -e config.bash ]; then
 fi
 
 source config.bash
-if [ -z "${SERVER}" -o -z "${SERVER_PORT}" -o -z "${KNOWN_HOSTS_STRING}" ]; then
-    message "Config variables are missing. export SERVER SERVER_PORT and KNOWN_HOSTS_STRING and run this setup again."
+if [ -z "${KNOWN_HOSTS_STRING}" -o -z "${BACKUP_DIRECTORY}" ]; then
+    message "Config variables are missing. export SERVER SERVER_PORT BACKUP_DIRECTORY and KNOWN_HOSTS_STRING and run this setup again."
+    exit 1
+fi
+if [ -z "${SERVER}" -a -z "${SERVER_PORT}" -a -z "${LOCAL_DESTINATION_DIRECTORY}" ]; then
+    message "Config variables are missing. Either set SERVER and SERVER_PORT or set LOCAL_DESTINATION_DIRECTORY."
     exit 1
 fi
 
-while ! (which sftp && which wget && which openssl) >/dev/null; do
-    message "Unable to find some executable (sftp, wget, openss), they need to be installed"
+if [ "${LOCAL_DESTINATION_DIRECTORY}" ] && [ ! -d "${LOCAL_DESTINATION_DIRECTORY}" ]; then
+    echo "${LOCAL_DESTINATION_DIRECTORY} directory doesn't exist"
+    exit 1
+fi
+
+while ! (which sftp && which wget && which openssl && which python3) >/dev/null; do
+    message "Unable to find some executable (sftp, wget, openssl, python3), they need to be installed"
     read -s -n 1 -p "Go do this, then press any key to continue . . ."
 done
 
-mkdir --parents --verbose ${DUPLICACY_BASEDIR}/bin ${DUPLICACY_BASEDIR}/keys ${DUPLICACY_BASEDIR}/backup ${DUPLICACY_BASEDIR}/logs
+mkdir --parents --verbose ${DUPLICACY_BASEDIR}/bin ${DUPLICACY_BASEDIR}/keys ${BACKUP_DIRECTORY} ${DUPLICACY_BASEDIR}/logs
 chmod --verbose 700 ${DUPLICACY_BASEDIR}/keys
 if [[ ! -e "${DUPLICACY_BASEDIR}/bin/duplicacy" ]]; then
     wget "$DUPLICACY_URL" -O "$DUPLICACY_BASEDIR/bin/duplicacy"
     chmod --verbose 755 ${DUPLICACY_BASEDIR}/bin/duplicacy
 fi
 
-while [ $(ls -1 ${DUPLICACY_BASEDIR}/backup | wc -l) -le 1 ]; do
-    message "You'll need to setup all you symbolic links in ${DUPLICACY_BASEDIR}/backup/"
-    message "Example : ln --verbose --symbolic /boot ${DUPLICACY_BASEDIR}/backup/boot"
+while [ $(ls -1 ${BACKUP_DIRECTORY} | wc -l) -lt 1 ]; do
+    message "You'll need to setup all you symbolic links in ${BACKUP_DIRECTORY}"
+    message "Example : ln --verbose --symbolic /boot ${BACKUP_DIRECTORY}/boot"
     read -s -n 1 -p "Go do this, then press any key to continue . . ."
 done
 
@@ -51,10 +60,12 @@ while [[ ! -e "${KEY_FILE}" ]]; do
 done
 chmod --verbose 600 ${KEY_FILE}
 
+# If the known_hosts file doesn't exist or the contents differ from the KNOWN_HOSTS_STRING, update the file
 KNOWN_HOSTS=${DUPLICACY_BASEDIR}/keys/known_hosts
-echo -e "${KNOWN_HOSTS_STRING}" > ${KNOWN_HOSTS}
-
-cd ${DUPLICACY_BASEDIR}/backup
+if [ ! -e "${KNOWN_HOSTS}" ] || [[ "$(< ${KNOWN_HOSTS})" != "${KNOWN_HOSTS_STRING}" ]]; then
+    echo -e "${KNOWN_HOSTS_STRING}" > ${KNOWN_HOSTS}
+fi
+cd ${BACKUP_DIRECTORY}
 
 while [ -z "$ENCRYPTION_ARGUMENT" ]; do
     read -p "Would you like synchronous encryption or RSA asynchronous encryption? (sync / async) : "
@@ -77,21 +88,34 @@ while [ -z "$ENCRYPTION_ARGUMENT" ]; do
     fi
 done
 
-message "About to initialize the backup. Enter the encryption password you'd like to use with this backup when prompted for the \"storage password\""
-DUPLICACY_SSH_KEY_FILE="${KEY_FILE}" "${DUPLICACY_BASEDIR}/bin/duplicacy" init ${ENCRYPTION_ARGUMENT} -repository ${DUPLICACY_BASEDIR}/backup ${CLIENT} sftp://${CLIENT}@${SERVER}:${SERVER_PORT}/backup
-# This is where you interactively enter the password
-echo "mkdir backup/logs" | sftp -o Port=${SERVER_PORT} -o IdentityFile=${KEY_FILE} -o BatchMode=yes -o UserKnownHostsFile=${KNOWN_HOSTS} ${CLIENT}@${SERVER}
+if [ "${SERVER}" -a "${SERVER_PORT}" ]; then
+    DESTINATION_URI="sftp://${CLIENT}@${SERVER}:${SERVER_PORT}/backup"
+elif [ "${LOCAL_DESTINATION_DIRECTORY}" ]; then
+    DESTINATION_URI="${LOCAL_DESTINATION_DIRECTORY}"
+else
+    echo "Missing either SERVER and SERVER_PORT or LOCAL_DESTINATION_DIRECTORY"
+    exit 1
+fi
 
-FILTER_FILE="${DUPLICACY_BASEDIR}/backup/.duplicacy/filters"
+message "About to initialize the backup. Enter the encryption password you'd like to use with this backup when prompted for the \"storage password\""
+DUPLICACY_SSH_KEY_FILE="${KEY_FILE}" "${DUPLICACY_BASEDIR}/bin/duplicacy" init ${ENCRYPTION_ARGUMENT} -repository ${BACKUP_DIRECTORY} ${CLIENT} ${DESTINATION_URI}
+# This is where you interactively enter the password
+if [ "${LOCAL_DESTINATION_DIRECTORY}" ]; then
+    mkdir --verbose "${LOCAL_DESTINATION_DIRECTORY}/logs"
+else
+    echo "mkdir backup/logs" | sftp -o Port=${SERVER_PORT} -o IdentityFile=${KEY_FILE} -o BatchMode=yes -o UserKnownHostsFile=${KNOWN_HOSTS} ${CLIENT}@${SERVER}
+fi
+
+FILTER_FILE="${BACKUP_DIRECTORY}/.duplicacy/filters"
 if [[ ! -e "${FILTER_FILE}" ]]; then
-    message "Fetching /opt/duplicacy/backup/.duplicacy/filters"
+    message "Fetching ${BACKUP_DIRECTORY}/.duplicacy/filters"
     wget -O "${FILTER_FILE}" https://raw.githubusercontent.com/gene1wood/personal-backup-system/master/duplicacy-filters-linux.txt
 fi
 
 "${DUPLICACY_BASEDIR}/bin/duplicacy" set -key ssh_key_file -value "${KEY_FILE}"
 
-while ! grep '"password"' ${DUPLICACY_BASEDIR}/backup/.duplicacy/preferences >/dev/null; do
-    message "You'll need to manually add the encryption password to /opt/duplicacy/backup/.duplicacy/preferences under keys... password"
+while ! grep '"password"' ${BACKUP_DIRECTORY}/.duplicacy/preferences >/dev/null; do
+    message "You'll need to manually add the encryption password to ${BACKUP_DIRECTORY}/.duplicacy/preferences under keys... password"
     # Probably want to avoid using set until this bug is fixed https://github.com/gilbertchen/duplicacy/issues/526
     # Or maybe not? Maybe the issue was the double quotes around the password? Though "&<>" are all escaped by the go json library
     # it doesn't seem to be a problem.
@@ -102,8 +126,20 @@ while ! grep '"password"' ${DUPLICACY_BASEDIR}/backup/.duplicacy/preferences >/d
 done
 
 if [[ ! -e "${DUPLICACY_BASEDIR}/bin/config.bash" ]]; then
-    echo "SERVER_PORT=${SERVER_PORT}" > "${DUPLICACY_BASEDIR}/bin/config.bash"
-    echo "SERVER=${SERVER}" >> "${DUPLICACY_BASEDIR}/bin/config.bash"
+    while [ -z "$hc_uuid" ]; do
+        read -p "What is this clients healthchecks.io UUID : "
+        if [[ $REPLY =~ ^\{?[A-F0-9a-f]{8}-[A-F0-9a-f]{4}-[A-F0-9a-f]{4}-[A-F0-9a-f]{4}-[A-F0-9a-f]{12}\}?$ ]]; then
+            hc_uuid=$REPLY
+        fi
+    done
+    echo "HC_UUID=${hc_uuid}" >> "${DUPLICACY_BASEDIR}/bin/config.bash"
+    echo "BACKUP_DIRECTORIES=(\"${BACKUP_DIRECTORY}\")" >> "${DUPLICACY_BASEDIR}/bin/config.bash"
+else
+    # Because programatically modifying the /opt/duplicacy/bin/config.bash file is super complex, just tell the user to do it
+    # This assumed that BACKUP_DIRECTORIES is set in ${DUPLICACY_BASEDIR}/bin/config.bash (which it might not be)
+
+    # TODO : Make this echo the green message color
+    bash -c "source \"${DUPLICACY_BASEDIR}/bin/config.bash\"; if ! printf '%s\0' \"\${BACKUP_DIRECTORIES[@]}\" | grep -Fxqz -- \"\$0\"; then echo -e \"Modify /opt/duplicacy/bin/config.bash and add \$0 to BACKUP_DIRECTORIES_ARRAY\"; fi" "${BACKUP_DIRECTORY}"
 fi
 
 if [[ ! -e "${DUPLICACY_BASEDIR}/bin/run-scheduled-duplicacy-backup.bash" ]]; then
@@ -111,10 +147,10 @@ if [[ ! -e "${DUPLICACY_BASEDIR}/bin/run-scheduled-duplicacy-backup.bash" ]]; th
     chmod --verbose 755 "${DUPLICACY_BASEDIR}/bin/run-scheduled-duplicacy-backup.bash"
 fi
 
-
 init_system=$(ps --no-headers -o comm 1)
 if [ "${init_system}" = "systemd" ]; then
-    cat << END-OF-FILE > /etc/systemd/system/scheduled-duplicacy-backup.service
+    if [ ! -e /etc/systemd/system/scheduled-duplicacy-backup.service ]; then
+        cat << END-OF-FILE > /etc/systemd/system/scheduled-duplicacy-backup.service
 [Unit]
 Description=Incremental backup with Duplicacy followed by pruning
 
@@ -122,12 +158,14 @@ Description=Incremental backup with Duplicacy followed by pruning
 Type=oneshot
 ExecStart=${DUPLICACY_BASEDIR}/bin/run-scheduled-duplicacy-backup.bash
 END-OF-FILE
-    cat << END-OF-FILE > /etc/systemd/system/scheduled-duplicacy-backup.timer
+    fi
+    if [ ! -e /etc/systemd/system/scheduled-duplicacy-backup.timer ]; then
+        cat << END-OF-FILE > /etc/systemd/system/scheduled-duplicacy-backup.timer
 [Unit]
-Description=Run scheduled-duplicacy-backup.service every night between 1AM and 4AM
+Description=Run scheduled-duplicacy-backup.service every night between 1AM and 4AM PST
 
 [Timer]
-OnCalendar=*-*-* 01:00:00
+OnCalendar=*-*-* 09:00:00 UTC
 # triggers the service immediately if it missed the last start time
 Persistent=true
 # Delay between 0 and 3 hours in seconds (3 * 60 * 60 = 10800)
@@ -136,26 +174,33 @@ RandomizedDelaySec=10800
 [Install]
 WantedBy=timers.target
 END-OF-FILE
-    systemctl enable scheduled-duplicacy-backup.timer --now
+        systemctl enable scheduled-duplicacy-backup.timer --now
+    fi
     message "# To do a real first backup"
     message "# For systemd run (yes, you actually need screen because this will run for a long time)"
     message "screen"
     message "systemctl start scheduled-duplicacy-backup.service"
 else
     # Note : the \$ is to escape the HEREDOC. The \% is to escape the crontab
-    cat << END-OF-FILE > /etc/cron.d/scheduled-duplicacy-backup.cron
+    if [ ! -e /etc/cron.d/scheduled-duplicacy-backup.cron ]; then
+        cat << END-OF-FILE > /etc/cron.d/scheduled-duplicacy-backup.cron
 SHELL=/bin/bash
 # Run scheduled-duplicacy-backup.service every night between 1AM and 4AM
 # Delay between 0 and 3 hours in seconds (3 * 60 * 60 = 10800)
 0 1 * * * root sleep \$[RANDOM \% 10800]; ${DUPLICACY_BASEDIR}/bin/run-scheduled-duplicacy-backup.bash
 END-OF-FILE
-  message "# To do a real first backup"
-  message "# For cron run"
-  message "screen"
-  message "/opt/duplicacy/bin/run-scheduled-duplicacy-backup.bash"
+    fi
+    message "# To do a real first backup"
+    message "# For cron run"
+    message "screen"
+    message "/opt/duplicacy/bin/run-scheduled-duplicacy-backup.bash"
 fi
 
 message "To do a dry run to see what would be backed up run"
-message "cd ${DUPLICACY_BASEDIR}/backup"
+message "cd ${BACKUP_DIRECTORY}"
 message "${DUPLICACY_BASEDIR}/bin/duplicacy backup -stats -dry-run | tee -a ${DUPLICACY_BASEDIR}/logs/duplicacy.dry-run.${CLIENT}.\`date +%Y%m%d%H%M%S\`.txt"
-message "And if you used RSA encryption, remove the private key from the client and store it somewhere safe"
+case $ENCRYPTION_ARGUMENT in
+    *-key* )
+         message "Make sure to remove the RSA private key from the client and store it somewhere safe"
+         ;;
+esac
